fixed some stuff, addd links, implemented the header and footer in the other pages
i added links between the person feed and add person pahes, the button didnt work, the path had to be tetween ' ' so i added them and now it works. From now on you should use the code from the header_and_footer.html file because it has the most recent changes and paths. i also added place holder paths for the pages that aren't done yet, so when you uploar your pages you should be able to just replace the place holder with a path to your page.